+++ packages/npm/lib/react/select.js
@@ -85,10 +85,8 @@ var Select = function (_React$Component) {
   }, {
     key: 'onInnerChange',
     value: function onInnerChange(ev) {
-      var value = ev.target.value;
-
       // update state
-      this.setState({ value: value });
+      this.setState({ value: ev.target.value });
     }
   }, {
     key: 'onInnerMouseDown',
@@ -156,7 +154,10 @@ var Select = function (_React$Component) {
     }
   }, {
     key: 'hideMenu',
-    value: function hideMenu() {
+    value: function hideMenu(ev) {
+      // check default prevented
+      if (ev && ev.defaultPrevented) return;
+
       // remove event listeners
       jqLite.off(window, 'resize', this.hideMenuCB);
       jqLite.off(document, 'click', this.hideMenuCB);
@@ -169,11 +170,11 @@ var Select = function (_React$Component) {
     }
   }, {
     key: 'onMenuChange',
-    value: function onMenuChange(value) {
+    value: function onMenuChange(index) {
       if (this.props.readOnly) return;
 
       // update inner <select> and dispatch 'change' event
-      this.controlEl.value = value;
+      this.controlEl.selectedIndex = index;
       util.dispatchEvent(this.controlEl, 'change');
     }
   }, {
@@ -181,7 +182,11 @@ var Select = function (_React$Component) {
     value: function render() {
       var _this2 = this;
 
-      var menuElem = void 0;
+      var value = this.state.value,
+          valueArgs = {},
+          menuElem = void 0,
+          placeholderElem = void 0,
+          selectCls = void 0;
 
       if (this.state.showMenu) {
         menuElem = _react2.default.createElement(Menu, {
@@ -211,8 +216,27 @@ var Select = function (_React$Component) {
           disabled = _props.disabled,
           useDefault = _props.useDefault,
           name = _props.name,
-          reactProps = babelHelpers.objectWithoutProperties(_props, ['children', 'className', 'style', 'label', 'defaultValue', 'readOnly', 'disabled', 'useDefault', 'name']);
+          placeholder = _props.placeholder,
+          reactProps = babelHelpers.objectWithoutProperties(_props, ['children', 'className', 'style', 'label', 'defaultValue', 'readOnly', 'disabled', 'useDefault', 'name', 'placeholder']);
 
+      // build value arguments
+
+      if (this.props.value !== undefined) valueArgs.value = value; // controlled
+      if (defaultValue !== undefined) valueArgs.defaultValue = defaultValue;
+
+      // handle placeholder
+      if (placeholder) {
+        placeholderElem = _react2.default.createElement(
+          'option',
+          { className: 'mui--text-placeholder', value: '' },
+          placeholder
+        );
+
+        // apply class if value is empty
+        if (value === '' || value === undefined && !defaultValue) {
+          selectCls = 'mui--text-placeholder';
+        }
+      }
 
       return _react2.default.createElement(
         'div',
@@ -228,20 +252,20 @@ var Select = function (_React$Component) {
         }),
         _react2.default.createElement(
           'select',
-          {
+          babelHelpers.extends({}, valueArgs, {
             ref: function ref(el) {
               _this2.controlEl = el;
             },
+            className: selectCls,
             name: name,
             disabled: disabled,
             tabIndex: tabIndexInner,
-            value: this.state.value,
-            defaultValue: defaultValue,
             readOnly: readOnly,
             onChange: this.onInnerChangeCB,
             onMouseDown: this.onInnerMouseDownCB,
             required: this.props.required
-          },
+          }),
+          placeholderElem,
           children
         ),
         _react2.default.createElement(
@@ -265,6 +289,7 @@ var Select = function (_React$Component) {
 Select.defaultProps = {
   className: '',
   name: '',
+  placeholder: null,
   readOnly: false,
   useDefault: typeof document !== 'undefined' && 'ontouchstart' in document.documentElement ? true : false,
   onChange: null,
@@ -282,7 +307,7 @@ var Menu = function (_React$Component2) {
 
     _this3.state = {
       origIndex: null,
-      currentIndex: null
+      currentIndex: 0
     };
 
 
@@ -290,21 +315,34 @@ var Menu = function (_React$Component2) {
     _this3.onKeyPressCB = util.callback(_this3, 'onKeyPress');
     _this3.q = '';
     _this3.qTimeout = null;
+    _this3.availOptionEls = [];
+
+    // extract selectable options
+    var optionEls = props.optionEls,
+        el = void 0,
+        i = void 0;
+
+    for (i = 0; i < optionEls.length; i++) {
+      el = optionEls[i];
+      if (!el.disabled && !el.hidden) _this3.availOptionEls.push(el);
+    }
     return _this3;
   }
 
   babelHelpers.createClass(Menu, [{
     key: 'componentWillMount',
     value: function componentWillMount() {
-      var optionEls = this.props.optionEls,
+      var optionEls = this.availOptionEls,
           m = optionEls.length,
-          selectedPos = 0,
+          selectedPos = null,
           i = void 0;
 
       // get current selected position
       for (i = m - 1; i > -1; i--) {
         if (optionEls[i].selected) selectedPos = i;
-      }this.setState({ origIndex: selectedPos, currentIndex: selectedPos });
+      }if (selectedPos !== null) {
+        this.setState({ origIndex: selectedPos, currentIndex: selectedPos });
+      }
     }
   }, {
     key: 'componentDidMount',
@@ -338,8 +376,9 @@ var Menu = function (_React$Component2) {
     key: 'onClick',
     value: function onClick(pos, ev) {
       // don't allow events to bubble
-      ev.stopPropagation();
-      this.selectAndDestroy(pos);
+      //ev.stopPropagation();
+      ev.preventDefault();
+      if (pos !== null) this.selectAndDestroy(pos);
     }
   }, {
     key: 'onKeyDown',
@@ -369,7 +408,7 @@ var Menu = function (_React$Component2) {
 
       // select first match alphabetically
       var prefixRegex = new RegExp('^' + this.q, 'i'),
-          optionEls = this.props.optionEls,
+          optionEls = this.availOptionEls,
           m = optionEls.length,
           i = void 0;
 
@@ -384,7 +423,7 @@ var Menu = function (_React$Component2) {
   }, {
     key: 'increment',
     value: function increment() {
-      if (this.state.currentIndex === this.props.optionEls.length - 1) return;
+      if (this.state.currentIndex === this.availOptionEls.length - 1) return;
       this.setState({ currentIndex: this.state.currentIndex + 1 });
     }
   }, {
@@ -400,7 +439,7 @@ var Menu = function (_React$Component2) {
 
       // handle onChange
       if (pos !== this.state.origIndex) {
-        this.props.onChange(this.props.optionEls[pos].value);
+        this.props.onChange(this.availOptionEls[pos].index);
       }
 
       // close menu
@@ -437,33 +476,51 @@ var Menu = function (_React$Component2) {
       var menuItems = [],
           optionEls = this.props.optionEls,
           m = optionEls.length,
+          pos = 0,
           optionEl = void 0,
           cls = void 0,
+          val = void 0,
           i = void 0;
 
       // define menu items
       for (i = 0; i < m; i++) {
-        cls = i === this.state.currentIndex ? 'mui--is-selected ' : '';
+        optionEl = optionEls[i];
+
+        // handle hidden
+        if (optionEl.hidden) continue;
+
+        // handle disabled
+        if (optionEl.disabled) {
+          cls = 'mui--is-disabled ';
+          val = null;
+        } else {
+          cls = pos === this.state.currentIndex ? 'mui--is-selected ' : '';
+          val = pos;
+          pos += 1;
+        }
 
         // add custom css class from <Option> component
-        cls += optionEls[i].className;
+        cls += optionEl.className;
 
         menuItems.push(_react2.default.createElement(
           'div',
           {
             key: i,
             className: cls,
-            onClick: this.onClick.bind(this, i)
+            onClick: this.onClick.bind(this, val)
           },
-          optionEls[i].textContent
+          optionEl.textContent
         ));
       }
 
       return _react2.default.createElement(
         'div',
-        { ref: function ref(el) {
+        {
+          ref: function ref(el) {
             _this4.wrapperElRef = el;
-          }, className: 'mui-select__menu' },
+          },
+          className: 'mui-select__menu'
+        },
         menuItems
       );
     }